pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "job-bot"
version = "0.1.0"
description = "Configurable automation bot that applies to jobs by filling online forms"
authors = [
  {name = "Automation Bot", email = "example@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "playwright>=1.40.0",
  "typer[all]>=0.9.0",
  "jinja2>=3.1.0"
]

[project.scripts]
job-bot = "job_bot.cli:app"
